Add foreign key to user chatbot; Add api for user chatbot login and get by id; Edit tenant and user entity

+++ src/controller/UserChatbotController.ts
@@ -1,6 +1,9 @@
 import { Request, Response, Router } from 'express';
 import { Logger } from '../utils/Logger';
-import { authorize, requireAuthentication } from '../middleware/auth.middleware';
+import {
+  authorize,
+  requireAuthentication,
+} from '../middleware/auth.middleware';
 import { UserChatbotDto } from '../dto/userChatbot.dto';
 import { UserChatbotService } from '../services/UserChatbotService';
 
@@ -11,7 +14,8 @@ const logger = Logger.getInstance();
 router.post('/create', async (req: Request, res: Response) => {
   try {
     const userChatbotDto: UserChatbotDto = req.body;
-    const result = await userChatbotService.createUserChatbotRequest(userChatbotDto);
+    const result =
+      await userChatbotService.createUserChatbotRequest(userChatbotDto);
     res.status(201).json(result);
   } catch (error) {
     console.error('Error creating user chatbot', error);
@@ -19,16 +23,31 @@ router.post('/create', async (req: Request, res: Response) => {
   }
 });
 
-
-router.delete('/delete',requireAuthentication, authorize(['SUPERADMIN', 'TENANT']), async (req: Request, res: Response) => {
+router.get('/:id', async (req: Request, res: Response) => {
   try {
-    const userChatbotDto: UserChatbotDto = req.body;
-    const result = await userChatbotService.deleteUserChatbot(userChatbotDto);
-    res.status(201).json(result);
+    const id = parseInt(req.params.id);
+    const result = await userChatbotService.getUserChatbotById(id);
+    res.status(200).json(result);
   } catch (error) {
-    console.error('Error deleting user chatbot', error);
+    console.error('Error fetching user chatbot', error);
     res.status(500).json({ error: 'internal server error' });
   }
-})
+});
+
+router.delete(
+  '/delete',
+  requireAuthentication,
+  authorize(['SUPERADMIN', 'TENANT']),
+  async (req: Request, res: Response) => {
+    try {
+      const userChatbotDto: UserChatbotDto = req.body;
+      const result = await userChatbotService.deleteUserChatbot(userChatbotDto);
+      res.status(201).json(result);
+    } catch (error) {
+      console.error('Error deleting user chatbot', error);
+      res.status(500).json({ error: 'internal server error' });
+    }
+  },
+);
 
 export default router;
